Fix rendering artifacts in video-7 mixed mode (pt.1)
Rewrote most of the mixed video-7 rendering code to properly handle
when the 'mode' changes in bytes that don't align with the 4-bit dhires
pixel color codes.

It resolves the artifacts I see in the AppleColor demo disk progams,
but is currently too slow to render the entire screen in time.