Fix CPU reset detection
This fixes the card inadvertently resetting and switching
to text mode between levels in Stellar 7.

I think this happens because Stellar 7 uses the ROMs as randomness for the
warp sounds and if it happened to read the reset vector address then it would
have triggered this bug.

Thanks to @dhgoss for reporting!