Remove read_cycle support
I'm removing support for the firmware to be able to put data on the data
bus in response to a device read cycle. The GPIO currently connected to
the Data transceiver direction will be hard-coded as an input pulled low
to preserve compatibility with the current (rev A) schematic.

I originally intended to use this to add Videx card emulation but have
abandoned that idea because:
 * There's not a lot of software that would use it
 * I don't think emulating both a ROM and I/O registers can be done with
   the current hardware setup anyway

In the future it would be possible to repurpose the pin as an input for
the bus SYNC signal, so maybe it will be possible to synchronize the VGA
refresh with the Apple II refresh when the card is installed in slot 7.