Make hires rendering more similar to dhires
Refactor the hires rendering to work more like the dhires rendering,
which uses an in-memory bitmap from LSB to MSB to represent the video
dots.

Updated the hires color pattern generation with better "interference"
colors, and added the rgb_filter_demo script to preview the result
without having to load onto the VGA card.